- moves writers to their own namespace
